+++ src/main/java/com/globalkinetic/practical_test/controllers/UserController.java
@@ -1,7 +1,10 @@
 package com.globalkinetic.practical_test.controllers;
 
+import com.globalkinetic.practical_test.dto.LoginRequestDTO;
+import com.globalkinetic.practical_test.dto.LoginResponseDTO;
 import com.globalkinetic.practical_test.dto.UserRequestDTO;
 import com.globalkinetic.practical_test.dto.UserResponseDTO;
+import com.globalkinetic.practical_test.services.AuthService;
 import com.globalkinetic.practical_test.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -18,6 +21,12 @@ import java.util.List;
 @RequestMapping("/api")
 public class UserController {
     private UserService userService;
+    private AuthService authService;
+
+    @Autowired
+    public void setAuthService(AuthService authService) {
+        this.authService = authService;
+    }
 
     @Autowired
     public void setUserService(UserService userService) {
@@ -33,5 +42,10 @@ public class UserController {
     @GetMapping("users")
     public ResponseEntity<UserResponseDTO> getAllUsers() {
         return ResponseEntity.ok(userService.getAllUsers());
+    }
+
+    @PostMapping("login")
+    public ResponseEntity<LoginResponseDTO> login(@RequestBody LoginRequestDTO loginReq) {
+        return ResponseEntity.ok(authService.login(loginReq));
     }
 }