+++ src/main/java/com/globalkinetic/practical_test/enums/Role.java
@@ -1,0 +1,8 @@
+package com.globalkinetic.practical_test.enums;
+
+/**
+ * @author Joesta
+ */
+public enum Role {
+    ADMIN, USER
+}