+++ src/main/java/com/globalkinetic/practical_test/mappers/UserMapper.java
@@ -5,7 +5,6 @@ import com.globalkinetic.practical_test.dto.UserRequestDTO;
 import com.globalkinetic.practical_test.dto.UserResponseDTO;
 import com.globalkinetic.practical_test.models.User;
 import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
 
 import java.util.List;
 
@@ -16,7 +15,9 @@ import java.util.List;
 @Mapper(componentModel = "spring")
 public interface UserMapper {
     User toUserEntity(UserRequestDTO userRequestDTO);
+
     List<UserDTO> toUserDTOList(List<User> users);
+
     default UserResponseDTO toUserResponseDTOList(List<User> users) {
         return new UserResponseDTO(toUserDTOList(users));
     }