+++ src/main/java/com/globalkinetic/practical_test/services/jwt/JwtAuthenticationFilter.java
@@ -1,0 +1,78 @@
+package com.globalkinetic.practical_test.services.jwt;
+
+
+import com.globalkinetic.practical_test.models.User;
+import com.globalkinetic.practical_test.repository.UserRepo;
+import com.globalkinetic.practical_test.services.UserDetailsServiceImpl;
+import jakarta.servlet.FilterChain;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.ApplicationContext;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
+import org.springframework.stereotype.Component;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import java.io.IOException;
+
+/**
+ * @author Joesta
+ */
+@Component
+public class JwtAuthenticationFilter extends OncePerRequestFilter {
+
+    private JwtService jwtService;
+    private UserRepo userRepo;
+    private ApplicationContext appContext;
+
+
+    @Autowired
+    public void setAppContext(ApplicationContext appContext) {
+        this.appContext = appContext;
+    }
+
+    @Autowired
+    public void setUserRepo(UserRepo userRepo) {
+        this.userRepo = userRepo;
+    }
+
+    @Autowired
+    public void setJwtService(JwtService jwtService) {
+        this.jwtService = jwtService;
+    }
+
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
+
+        String authorizationHeader = request.getHeader("Authorization");
+        String token = null;
+        String username = null;
+
+        if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
+            token = authorizationHeader.substring(7);
+
+           try {
+               username = jwtService.getUsername(token);
+           } catch (RuntimeException re) {
+               throw new ServletException("Invalid JWT token", re);
+           }
+
+            User user = userRepo.findByUsername(username);
+
+            if (user != null && SecurityContextHolder.getContext().getAuthentication() == null) {
+                UserDetails userDetails = appContext.getBean(UserDetailsServiceImpl.class).loadUserByUsername(username);
+                if (jwtService.validateToken(token, userDetails)) {
+                    UsernamePasswordAuthenticationToken authToken = new UsernamePasswordAuthenticationToken(userDetails, null , userDetails.getAuthorities());
+                    authToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
+                    SecurityContextHolder.getContext().setAuthentication(authToken);
+                }
+            }
+        }
+
+        filterChain.doFilter(request, response);
+    }
+}